Using & validating the token | Adding auth middleware to all routes | connecting posts & users | Ading authorization checks | Clearing post-user realtion | Removing error message from frontend

+++ controllers/feed.js
@@ -1,13 +1,17 @@
 const fs = require('fs');
 const path = require('path');
+
 const { validationResult } = require('express-validator/check');
+
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
     const perPage = 2;
     let totalItems;
-    Post.find().countDocuments()
+    Post.find()
+        .countDocuments()
         .then(count => {
             totalItems = count;
             return Post.find()
@@ -15,54 +19,10 @@ exports.getPosts = (req, res, next) => {
                 .limit(perPage);
         })
         .then(posts => {
-            res
-                .status(200)
-                .json({
-                    message: 'Fetched posts successfully!',
-                    posts: posts,
-                    totalItems: totalItems
-                });
-        })
-        .catch(err => {
-            if (!err.statusCode) {
-                err.statusCode = 500;
-            }
-            next(err);
-        });
-
-};
-
-exports.createPost = (req, res, next) => {
-    const errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-        const error = new Error('Validation failed, entered data is incorrect!');
-        error.statusCode = 422;
-        throw error;
-    }
-    if (!req.file) {
-        const error = new Error('No image provided!');
-        error.statusCode = 422;
-        throw error;
-    }
-
-    const imageUrl = req.file.path.replace("\\", "/");
-    const title = req.body.title;
-    const content = req.body.content;
-    const post = new Post({
-        title: title,
-        content: content,
-        imageUrl: imageUrl,
-        creator: {
-            name: 'Merima'
-        }
-    });
-    post.save()
-        .then(result => {
-            console.log(result);
-            res.status(201).json({
-                message: 'Post created successfully!',
-                post: result
+            res.status(200).json({
+                message: 'Fetched posts successfully.',
+                posts: posts,
+                totalItems: totalItems
             });
         })
         .catch(err => {
@@ -71,7 +31,53 @@ exports.createPost = (req, res, next) => {
             }
             next(err);
         });
+};
 
+exports.createPost = (req, res, next) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        const error = new Error('Validation failed, entered data is incorrect.');
+        error.statusCode = 422;
+        throw error;
+    }
+    if (!req.file) {
+        const error = new Error('No image provided.');
+        error.statusCode = 422;
+        throw error;
+    }
+    const imageUrl = req.file.path.replace("\\", "/");
+    const title = req.body.title;
+    const content = req.body.content;
+    let creator;
+    const post = new Post({
+        title: title,
+        content: content,
+        imageUrl: imageUrl,
+        creator: req.userId
+    });
+    post
+        .save()
+        .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then(result => {
+            res.status(201).json({
+                message: 'Post created successfully!',
+                post: post,
+                creator: { _id: creator._id, name: creator.name }
+            });
+        })
+        .catch(err => {
+            if (!err.statusCode) {
+                err.statusCode = 500;
+            }
+            next(err);
+        });
 };
 
 exports.getPost = (req, res, next) => {
@@ -79,11 +85,11 @@ exports.getPost = (req, res, next) => {
     Post.findById(postId)
         .then(post => {
             if (!post) {
-                const error = new Error('No Post found in the database!');
+                const error = new Error('Could not find post.');
                 error.statusCode = 404;
                 throw error;
             }
-            return res.status(200).json({ message: 'Post fetched!', post: post });
+            res.status(200).json({ message: 'Post fetched.', post: post });
         })
         .catch(err => {
             if (!err.statusCode) {
@@ -94,34 +100,34 @@ exports.getPost = (req, res, next) => {
 };
 
 exports.updatePost = (req, res, next) => {
-
     const postId = req.params.postId;
     const errors = validationResult(req);
-
     if (!errors.isEmpty()) {
-        const error = new Error('Validation failed, entered data is incorrect!');
+        const error = new Error('Validation failed, entered data is incorrect.');
         error.statusCode = 422;
         throw error;
     }
     const title = req.body.title;
     const content = req.body.content;
     let imageUrl = req.body.image;
-
     if (req.file) {
         imageUrl = req.file.path.replace("\\", "/");
     }
-
     if (!imageUrl) {
-        const error = new Error('Np file picked!');
+        const error = new Error('No file picked.');
         error.statusCode = 422;
         throw error;
     }
-
     Post.findById(postId)
         .then(post => {
             if (!post) {
-                const error = new Error('No Post found in the database!');
+                const error = new Error('Could not find post.');
                 error.statusCode = 404;
+                throw error;
+            }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
                 throw error;
             }
             if (imageUrl !== post.imageUrl) {
@@ -133,10 +139,7 @@ exports.updatePost = (req, res, next) => {
             return post.save();
         })
         .then(result => {
-            res.status(200).json({
-                message: 'Post updated!',
-                post: result
-            })
+            res.status(200).json({ message: 'Post updated!', post: result });
         })
         .catch(err => {
             if (!err.statusCode) {
@@ -147,24 +150,32 @@ exports.updatePost = (req, res, next) => {
 };
 
 exports.deletePost = (req, res, next) => {
-
     const postId = req.params.postId;
     Post.findById(postId)
         .then(post => {
-            //Check logged in user later...
             if (!post) {
-                const error = new Error('No Post found in the database!');
+                const error = new Error('Could not find post.');
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
+            // Check logged in user
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
-            console.log(result);
-            res.status(200).json({
-                message: 'Post deleted!'
-            })
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
+            res.status(200).json({ message: 'Deleted post.' });
         })
         .catch(err => {
             if (!err.statusCode) {
@@ -176,9 +187,5 @@ exports.deletePost = (req, res, next) => {
 
 const clearImage = filePath => {
     filePath = path.join(__dirname, '..', filePath);
-    fs.unlink(filePath, err => {
-        console.log(err);
-    });
+    fs.unlink(filePath, err => console.log(err));
 };
-
-
